requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
timedelta==2020.12.3
uuid==1.30
